connect to pub/sub pt.2, h2 database and implement order api

+++ src/main/kotlin/io/kraftsman/gcp/contracts/NotifierMessageContract.kt
@@ -1,6 +1,6 @@
 package io.kraftsman.gcp.contracts
 
 interface NotifierMessageContract {
-    val orderId: Int
+    val orderId: Long
     val message: String
 }
